[code] implementation of different utils
